src/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
appnope==0.1.3
asgiref==3.6.0
astroid==2.12.13
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
banana-dev==4.0.2
bandit==1.7.4
billiard==3.6.4.0
black==22.12.0
boto3==1.26.114
botocore==1.29.114
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==15.0.1
constantly==15.1.0
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==3.6
cryptography==40.0.2
cycler==0.11.0
daphne==4.0.0
decorator==5.1.1
dill==0.3.6
Django==4.2
django-cors-headers==3.13.0
django-filter==22.1
django-guardian==2.4.0
django-prometheus==2.2.0
django-storages==1.13.2
django-tenants==3.4.7
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-guardian==0.3.0
djangorestframework-simplejwt==5.2.2
drf-exceptions-hog==0.3.0
drf-yasg==1.21.4
exceptiongroup==1.1.0
executing==1.2.0
factory-boy==3.2.1
Faker==15.3.4
flake8==6.0.0
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
gunicorn==20.1.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imageio==2.31.1
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
ipython==8.14.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
kombu==5.2.4
lazy-object-proxy==1.8.0
lazy_loader==0.2
llvmlite==0.40.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==0.4.3
networkx==3.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
openai==0.27.4
openapi==1.1.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==22.0
parso==0.8.3
pathspec==0.10.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
pluggy==1.0.0
pooch==1.7.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.23.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.6.0
pylint==2.15.8
PyMatting==1.1.8
pyOpenSSL==23.1.1
pyparsing==3.1.0
pyrsistent==0.19.2
pytest==7.2.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
redis==4.5.4
rembg==2.0.47
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.0
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
stack-data==0.6.2
stevedore==4.1.1
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.4.12
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.65.0
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.2
zope.interface==6.0
